+++ goodreads_sync/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Goodreads Sync Change Log
 
-## [1.16.0] - 2022-10-XX
+## [1.16.0] - 2022-10-16
 _All kiwidude plugins updated/migrated to: https://github.com/kiwidude68/calibre_plugins_
 ### Changed
 - **Breaking:** Drop PyQt4 support, require calibre 2.x or later.
@@ -8,7 +8,10 @@ _All kiwidude plugins updated/migrated to: https://github.com/kiwidude68/calibre
 - Removed help file, point to [GitHub Wiki](https://github.com/kiwidude68/calibre_plugins/wiki/Goodreads-Sync)
 - Replace Help link on configuration dialog with a button.
 ### Fixed
-- Various bugs related to the "Update reading progress" feature.
+- Update reading progress should only apply "read" shelf actions when progress >= 100 ([#2][i2])
+- Update reading progress should apply "currently-reading" shelf actions when progress < 100 ([#2][i2])
+- Disable the rating/review/date-read cehckboxes on the 'currently-reading' shelf. The 'read' shelf actions are used for these instead when you finish a book for reading progress purposes.
+- Various other bugs related to the "Update reading progress" feature.
 - Deleting menu rows could have wrong selected items. (@capink)
 
 ## [1.15.6] - 2022-08-15
@@ -341,3 +344,5 @@ _All kiwidude plugins updated/migrated to: https://github.com/kiwidude68/calibre
 
 ## [1.0.0] - 2011-02-27
 _Initial release of Goodreads Sync plugin_
+
+[i2]: https://github.com/kiwidude68/calibre_plugins/issues/2